[sw] Add new enhanced PMP API (work in progress)
The new API introduces a staging area, `epmp_state_t`, that is used
to separate PMP entry calculations from any CSR updates. PMP entry
configurations are written individually to `epmp_state_t` and then
the CSRs can be updated by calling `epmp_set`.

Separating the computation of the CSR values from any updates to the
actual CSRs allows the user of this API to maintain one or more
copies of the expected CSR state in either RAM or ROM to use for
validation purposes. Alternatively the user could decide to
recreate the expected state at specific checkpoints in their code
as needed.

The biggest issue with the approach to updating CSRs taken in this
patch is that we write to every PMP CSR when performing an update.
This could be a performance issue, although we anticipate modifying
the CSRs infrequently during execution of the mask ROM and ROM_EXT.
It does also however mean that there may be more places in the code
where a CSR could be overwritten incorrectly due to a bug or an
attack.

Unlike the existing PMP API we specify all region types the same
way using a struct containing `start` and `end` addresses. This
makes it easy to be explicit about the region boundaries as shown
in the example below.

The new EPMP API assumes that most `mseccfg` updates will occur in
assembly and that the API will only be used to update the Rule
Locking Bypass (`msecffg.RLB`) feature. In particular this API
assumes that Machine Mode Lockdown (`mseccfg.MML`) will be
disabled (though that can be verified by a call to `epmp_check`.)

The API has been designed so that Machine Mode Lockdown support
can be added at a later date if required by adding an additional
set of `Lockdown` permission enum values. It would make the API
clunkier though.

Example (configure, set and check using the same state):

  ```
  epmp_state_t state = {0};
  rese = epmp_toggle_rule_locking_bypass(&state,
             kEpmpRlbToggleEnabled);
  // check rse

  res0 = epmp_entry_configure_na4(&state, 0,
             (epmp_region_t){ .start = 0x0, .end = 0x4 },
             kEpmpPermissionsLockedMachineReadWriteUserReadWrite);
  res1 = epmp_entry_configure_napot(&state, 1,
             (epmp_region_t){ .start = 0x40, .end = 0x48 },
             kEpmpPermissionsLockedMachineReadExecuteUserReadExecute);
  // pmpaddr2 used as base address for entry 3
  res3 = epmp_entry_configure_tor(&state, 3,
             (epmp_region_t){ .start = 0x100, .end = 0x200 },
             kEpmpPermissionsLockedMachineReadWriteUserReadWrite);
  res4 = epmp_entry_configure_tor(&state, 4,
             (epmp_region_t){ .start = 0x200, .end = 0x300 },
             kEpmpPermissionsLockedMachineReadUserRead);
  // check res0, res1, res3 and res4

  ress = epmp_set(&state); // commit state to CSRs
  // check ress

  // ... later ...

  resc = epmp_check(&state); // validate state against CSRs
  // check resc

  ```