[sw] Add new enhanced PMP API (work in progress)
The new API introduces a staging area, `epmp_state_t`, that is used
to separate PMP entry calculations from any CSR updates. PMP entry
configurations are written individually to `epmp_state_t` and then
the CSRs can be updated by calling `epmp_set`.

There are advantages and disadvantages to this approach. The
biggest issue is that we write to every PMP CSR when performing an
update. This could be a performance issue, although we anticipate
modifying the CSRs infrequently during execution of the mask ROM and
ROM_EXT. It does also however mean that there may be more places in
the code where a CSR could be overwritten incorrectly due to a bug
or an attack.

Unlike the existing PMP API we specify all region types the same
way using a struct containing `start` and `end` addresses. This
makes it easy to be explicit about the region boundaries as shown
in the example below.

The new EPMP API assumes that `mseccfg` updates will occur in
assembly and that the API will only be used when `mseccfg.RLB`
is set and `mseccfg.MLL` is unset. The `mseccfg.MMWP` register
is also assumed to be set but this does not affect the API. We
could add support for `mseccfg.MLL` and `mseccfg.RLB` modes
by extending the number of possible permission values. If we do
we could extend `epmp_state_t` to include `mseccfg` but it may
be harder to use because of ordering issues.

Example (configure, set and check using the same state):

  ```
  epmp_state_t state = {0};
  res0 = epmp_entry_configure_na4(&state, 0,
             (epmp_region_t){ .start = 0x0, .end = 0x4 },
             kEpmpPermissionsUnlockedMachineAllUserReadWrite);
  res1 = epmp_entry_configure_napot(&state, 1,
             (epmp_region_t){ .start = 0x40, .end = 0x48 },
             kEpmpPermissionsUnlockedMachineAllUserReadExecute);
  res3 = epmp_entry_configure_tor(&state, 3,
             (epmp_region_t){ .start = 0x100, .end = 0x200 },
             kEpmpPermissionsUnlockedMachineAllUserReadWrite);
  res4 = epmp_entry_configure_tor(&state, 4,
             (epmp_region_t){ .start = 0x200, .end = 0x300 },
             kEpmpPermissionsUnlockedMachineAllUserRead);
  // check res0, res1, res3 and res4

  ress = epmp_set(&state); // commit state to CSRs
  // check ress

  // ... later ...

  resc = epmp_check(&state); // validate state against CSRs

  ```