[mask_rom] Implement memory protection module.
This is an implementation of the memory protection specification
with most configuration done in assembly. A C API is also provided
to make the ROM_EXT image executable.

A shadow copy of the PMP configuration registers is maintained in
the mask ROM and may be checked against the hardware register
values using the `epmp_state_check` function.

A note on test status:

The ePMP configuration for the mask ROM prevents access to the DV
address space. Tests write to the DV address space to signal their
status. Therefore all tests booted from the mask ROM will fail
unless they modify the ePMP configuration.

For the time being this ePMP reconfiguration is performed in the
ROM_EXT so that the tests continue to pass without modification.
This is intended to be a temporary measure and this functionality
should be absorbed into one of the test libraries.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>